Cambios en validacion para el log de precios

+++ shopify_ept/models/shopify_product_ept.py
@@ -271,20 +271,21 @@ class ShopifyProductProductEpt(models.Model):
             for variant in template.shopify_product_ids:
                 variant_vals = self.shopify_prepare_variant_vals(instance, template, variant, is_set_price,
                                                             is_set_basic_detail)
-                if variant_vals.get('price') > 15.00: 
-                    variants.append(variant_vals)
-                    product_in_log_exist = self.env['product.log'].search([('product_id', '=', variant.product_id.id), ('instance_id', '=', instance.id)])
-                    if product_in_log_exist: 
-                        product_in_log_exist.unlink()
-                else: 
-                    product_in_log_exist = self.env['product.log'].search([('product_id', '=', variant.product_id.id), ('instance_id', '=', instance.id)])
-                    if not product_in_log_exist: 
-                        create_product_in_log = self.env['product.log'].create(
-                            {
-                                'product_id' : variant.product_id.id, 
-                                'instance_id': instance.id
-                            }
-                        )
+                if variant_vals.get('price') != None: 
+                    if variant_vals.get('price') > 15.00: 
+                        variants.append(variant_vals)
+                        product_in_log_exist = self.env['product.log'].search([('product_id', '=', variant.product_id.id), ('instance_id', '=', instance.id)])
+                        if product_in_log_exist: 
+                            product_in_log_exist.unlink()
+                    else: 
+                        product_in_log_exist = self.env['product.log'].search([('product_id', '=', variant.product_id.id), ('instance_id', '=', instance.id)])
+                        if not product_in_log_exist: 
+                            create_product_in_log = self.env['product.log'].create(
+                                {
+                                    'product_id' : variant.product_id.id, 
+                                    'instance_id': instance.id
+                                }
+                            )
                 
             new_product.variants = variants
         if is_set_basic_detail:
